+++ src/ui/index.js
@@ -1,1 +1,3 @@
+export { AppShell } from "./AppShell.js";
 export { ControlPanel } from "./ControlPanel.js";
+export { VisualSettingsPanel } from "./VisualSettingsPanel.js";